requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
idna==3.6
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
